refactor(build): hoist cosine_similarity from me-backend-sqlite to me-types
cosine_similarity is pure math (no SQL, no rusqlite, no driver
dependency) that ten call sites across four layers already consume.
Leaving it in me-backend-sqlite meant any primitive that needed it
(the forthcoming me-archive crate, an L3 leaf) would have to depend
on a concrete storage backend, dragging rusqlite+bundled-SQLite into
a Postgres-only build.

Moves the fn (+ its unit tests + the proptest_cosine property suite)
into me-types/src/math.rs. me-backend-sqlite re-exports it so its
internal callers (search/{vector,ann,mod}.rs) and the existing path
me_backend_sqlite::search::vector::cosine_similarity keep resolving;
the facade's memory_engine::search::cosine_similarity public path is
unchanged.

Also splits two doc-comment first paragraphs in me-types (introduced
by the prior MemoryQuery hoist) that clippy::too_long_first_doc_paragraph
only started flagging once this commit's --all-features recompile
picked them up.

Pure relocation: no behavior change, no public API change.
